chore: rename MobileNav to CompactNav to make it more behavioural
