+++ frontend/src/components/layout/CompactNav.tsx
